feat: setup products pagination

+++ database/migrations/2024_01_07_080602_create_products_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('price');
+            $table->unsignedInteger('price');
             $table->string('unit');
             $table->timestamps();
         });
